feat: refactor ssh
