+++ src/ParamikoMock/exceptions.py
@@ -1,0 +1,2 @@
+class BadSetupError(Exception):
+    pass
